feature: add schema to envVar and refactor the service data structure
