+++ index.js
@@ -1,5 +1,5 @@
 import {app} from './src/server.js'
 
-app.listen(3000, () => {
-    console.info("Server is runner in port 3000")
+app.listen(process.env.SERVICE_PORT, () => {
+    console.info(`Server is runner in port ${process.env.SERVICE_PORT}`)
 })
