+++ package.json
@@ -12,9 +12,11 @@
   },
   "dependencies": {
     "body-parser": "^1.19.0",
+    "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "mongodb": "^4.1.3",
-    "nodemon": "^2.0.13"
+    "nodemon": "^2.0.13",
+    "schema-js": "^1.1.0"
   }
 }
