+++ src/controller/authenticate/index.js
@@ -1,0 +1,3 @@
+import {auth} from "./routers/auth.js"
+
+export const authController = auth
