+++ src/controller/authenticate/routers/index.js
@@ -1,0 +1,2 @@
+import express from 'express';
+export const authenticate = express();
