+++ src/database/db.js
@@ -1,6 +1,9 @@
 import { default as mongodb } from 'mongodb';
+import {config} from '../config/index.js'
+
 let MongoClient = mongodb.MongoClient;
-export let url = "mongodb://localhost:27017/gameAPI";
+
+export let url = `mongodb://${config.database.host}:${config.database.port}/${config.database.name}`;
 
 MongoClient.connect(url, function (err, db) {
     if (err) throw err;
